aiu refactored to spyre
